address review: bypass list on Loader path, no Box::leak, delay-load winhttp
- SystemProxy::resolve now checks its own no_proxy list so the
  Loader::get_http_proxy_for path (install/upgrade/create/S3) honours the
  WinINet bypass list, not just the fetch path.
- Extract no_proxy matching into env_loader::no_proxy_list_matches and
  drop the duplicate copy in http/lib.rs.
- has_http_proxy() only counts a static system proxy, not auto-detect
  (which is on by default on stock Windows).
- Replace Box::leak in the PAC cache with an owned HashMap value plus a
  pointer-derived 'static borrow (PORTING.md forbids Box::leak).
- Drop the cfg(test) unit module: cargo test -p bun_dotenv doesn't link
  (highway C++ deps), and the integration test covers parsing via the
  env hook.
- Delay-load winhttp.dll in release; add it to the PE import allowlist.
